fix(menuBar): refactor code
