+++ src/components/Menu/components/DropdownMenu/interface.ts
@@ -1,0 +1,8 @@
+import * as React from 'react'
+
+export interface DropdownMenuProps {
+  label: string
+  nested?: boolean
+  children?: React.ReactNode
+  className?: string
+}
