+++ src/components/Menu/components/DropdownMenuContext/DropdownMenuContext.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react'
+
+export const DropdownMenuContext = React.createContext<{
+  getItemProps: (
+    userProps?: React.HTMLProps<HTMLElement>
+  ) => Record<string, unknown>
+  activeIndex: number | null
+  setActiveIndex: React.Dispatch<React.SetStateAction<number | null>>
+  setHasFocusInside: React.Dispatch<React.SetStateAction<boolean>>
+  isOpen: boolean
+}>({
+  getItemProps: () => ({}),
+  activeIndex: null,
+  setActiveIndex: () => {},
+  setHasFocusInside: () => {},
+  isOpen: false,
+})
