+++ src/components/Menu/components/DropdownMenuItem/interface.ts
@@ -1,0 +1,7 @@
+import * as React from 'react'
+
+export interface DropdownMenuItemProps {
+  children: React.ReactNode
+  disabled?: boolean
+  className?: string
+}
